moved serialization traits into their own module
