+++ utils/core/src/lib.rs
@@ -12,15 +12,15 @@
 #[macro_use]
 extern crate alloc;
 
+pub mod collections;
+pub mod iterators;
+pub mod string;
+
+use collections::Vec;
 use core::{convert::TryInto, mem, slice};
 
-pub mod collections;
-use collections::Vec;
-
-pub mod string;
-use string::ToString;
-
-pub mod iterators;
+mod serde;
+pub use serde::{ByteReader, ByteWriter, Deserializable, Serializable, SliceReader};
 
 mod errors;
 pub use errors::DeserializationError;
@@ -33,365 +33,6 @@ use iterators::*;
 
 #[cfg(feature = "concurrent")]
 pub use rayon;
-
-// SERIALIZABLE
-// ================================================================================================
-
-/// Defines how to serialize `Self` into bytes.
-pub trait Serializable: Sized {
-    // REQUIRED METHODS
-    // --------------------------------------------------------------------------------------------
-    /// Serializes `self` into bytes and writes these bytes into the `target`.
-    fn write_into<W: ByteWriter>(&self, target: &mut W);
-
-    // PROVIDED METHODS
-    // --------------------------------------------------------------------------------------------
-
-    /// Serializes `self` into a vector of bytes.
-    fn to_bytes(&self) -> Vec<u8> {
-        let mut result = Vec::with_capacity(self.get_size_hint());
-        self.write_into(&mut result);
-        result
-    }
-
-    /// Serializes all elements of the `source` and writes these bytes into the `target`.
-    ///
-    /// This method does not write any metadata (e.g. number of serialized elements) into the
-    /// `target`.
-    fn write_batch_into<W: ByteWriter>(source: &[Self], target: &mut W) {
-        for item in source {
-            item.write_into(target);
-        }
-    }
-
-    /// Returns an estimate of how many bytes are needed to represent self.
-    ///
-    /// The default implementation returns zero.
-    fn get_size_hint(&self) -> usize {
-        0
-    }
-}
-
-impl Serializable for () {
-    fn write_into<W: ByteWriter>(&self, _target: &mut W) {}
-}
-
-impl<T: Serializable> Serializable for Vec<T> {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        T::write_batch_into(self, target);
-    }
-}
-
-impl<T: Serializable> Serializable for &Vec<T> {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        T::write_batch_into(self, target);
-    }
-}
-
-impl<T: Serializable, const N: usize> Serializable for Vec<[T; N]> {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        let source = flatten_slice_elements(self);
-        T::write_batch_into(source, target);
-    }
-}
-
-impl<T: Serializable, const N: usize> Serializable for &Vec<[T; N]> {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        let source = flatten_slice_elements(self);
-        T::write_batch_into(source, target);
-    }
-}
-
-impl<T: Serializable> Serializable for &[T] {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        T::write_batch_into(self, target);
-    }
-}
-
-impl<T: Serializable, const N: usize> Serializable for &[[T; N]] {
-    fn write_into<W: ByteWriter>(&self, target: &mut W) {
-        let source = flatten_slice_elements(self);
-        T::write_batch_into(source, target);
-    }
-}
-
-// DESERIALIZABLE
-// ================================================================================================
-
-/// Defines how to deserialize `Self` from bytes.
-pub trait Deserializable: Sized {
-    // REQUIRED METHODS
-    // --------------------------------------------------------------------------------------------
-
-    /// Reads a sequence of bytes from the provided `source`, attempts to deserialize these bytes
-    /// into `Self`, and returns the result.
-    ///
-    /// # Errors
-    /// Returns an error if:
-    /// * The `source` does not contain enough bytes to deserialize `Self`.
-    /// * Bytes read from the `source` do not represent a valid value for `Self`.
-    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError>;
-
-    // PROVIDED METHODS
-    // --------------------------------------------------------------------------------------------
-
-    /// Reads a sequence of bytes from the provided `source`, attempts to deserialize these bytes
-    /// into a vector with the specified number of `Self` elements, and returns the result.
-    ///
-    /// # Errors
-    /// Returns an error if:
-    /// * The `source` does not contain enough bytes to deserialize the specified number of
-    ///   elements.
-    /// * Bytes read from the `source` do not represent a valid value for `Self` for any of the
-    ///   elements.
-    ///
-    /// Note: if the error occurs, the reader is not rolled back to the state prior to calling
-    /// this function.
-    fn read_batch_from<R: ByteReader>(
-        source: &mut R,
-        num_elements: usize,
-    ) -> Result<Vec<Self>, DeserializationError> {
-        let mut result = Vec::new();
-        for _ in 0..num_elements {
-            let element = Self::read_from(source)?;
-            result.push(element)
-        }
-        Ok(result)
-    }
-}
-
-// BYTE READER
-// ================================================================================================
-
-/// Defines how primitive values are to be read from `Self`.
-pub trait ByteReader {
-    /// Returns a single byte read from `self`.
-    ///
-    /// # Errors
-    /// Returns a [DeserializationError] error the reader is at EOF.
-    fn read_u8(&mut self) -> Result<u8, DeserializationError>;
-
-    /// Returns a u16 value read from `self` in little-endian byte order.
-    ///
-    /// # Errors
-    /// Returns a [DeserializationError] if a u16 value could not be read from `self`.
-    fn read_u16(&mut self) -> Result<u16, DeserializationError>;
-
-    /// Returns a u32 value read from `self` in little-endian byte order.
-    ///
-    /// # Errors
-    /// Returns a [DeserializationError] if a u32 value could not be read from `self`.
-    fn read_u32(&mut self) -> Result<u32, DeserializationError>;
-
-    /// Returns a u64 value read from `self` in little-endian byte order.
-    ///
-    /// # Errors
-    /// Returns a [DeserializationError] if a u64 value could not be read from `self`.
-    fn read_u64(&mut self) -> Result<u64, DeserializationError>;
-
-    /// Returns a u128 value read from `self` in little-endian byte order.
-    ///
-    /// # Errors
-    /// Returns a [DeserializationError] if a u128 value could not be read from `self`.
-    fn read_u128(&mut self) -> Result<u128, DeserializationError>;
-
-    /// Returns a byte vector of the specified length read from `self`.
-    ///
-    /// # Errors
-    /// Returns a [DeserializationError] if a vector of the specified length could not be read
-    /// from `self`.
-    fn read_u8_vec(&mut self, len: usize) -> Result<Vec<u8>, DeserializationError>;
-
-    /// Returns a byte array of length `N` reade from `self`.
-    fn read_u8_array<const N: usize>(&mut self) -> Result<[u8; N], DeserializationError>;
-
-    /// Returns true if there are more bytes left to be read from `self`.
-    fn has_more_bytes(&self) -> bool;
-}
-
-// SLICE READER
-// ================================================================================================
-
-/// Implements [ByteReader] trait for a slice of bytes.
-pub struct SliceReader<'a> {
-    source: &'a [u8],
-    pos: usize,
-}
-
-impl<'a> SliceReader<'a> {
-    /// Creates a new slice reader from the specified slice.
-    pub fn new(source: &'a [u8]) -> Self {
-        SliceReader { source, pos: 0 }
-    }
-}
-
-impl<'a> ByteReader for SliceReader<'a> {
-    fn read_u8(&mut self) -> Result<u8, DeserializationError> {
-        if self.pos >= self.source.len() {
-            return Err(DeserializationError::UnexpectedEOF);
-        }
-        let result = self.source[self.pos];
-
-        self.pos += 1;
-        Ok(result)
-    }
-
-    fn read_u16(&mut self) -> Result<u16, DeserializationError> {
-        let end_pos = self.pos + 2;
-        if end_pos > self.source.len() {
-            return Err(DeserializationError::UnexpectedEOF);
-        }
-
-        let result = u16::from_le_bytes(
-            self.source[self.pos..end_pos]
-                .try_into()
-                .map_err(|err| DeserializationError::UnknownError(format!("{err}")))?,
-        );
-
-        self.pos = end_pos;
-        Ok(result)
-    }
-
-    fn read_u32(&mut self) -> Result<u32, DeserializationError> {
-        let end_pos = self.pos + 4;
-        if end_pos > self.source.len() {
-            return Err(DeserializationError::UnexpectedEOF);
-        }
-
-        let result = u32::from_le_bytes(
-            self.source[self.pos..end_pos]
-                .try_into()
-                .map_err(|err| DeserializationError::UnknownError(format!("{err}")))?,
-        );
-
-        self.pos = end_pos;
-        Ok(result)
-    }
-
-    fn read_u64(&mut self) -> Result<u64, DeserializationError> {
-        let end_pos = self.pos + 8;
-        if end_pos > self.source.len() {
-            return Err(DeserializationError::UnexpectedEOF);
-        }
-
-        let result = u64::from_le_bytes(
-            self.source[self.pos..end_pos]
-                .try_into()
-                .map_err(|err| DeserializationError::UnknownError(format!("{err}")))?,
-        );
-
-        self.pos = end_pos;
-        Ok(result)
-    }
-
-    fn read_u128(&mut self) -> Result<u128, DeserializationError> {
-        let end_pos = self.pos + 16;
-        if end_pos > self.source.len() {
-            return Err(DeserializationError::UnexpectedEOF);
-        }
-
-        let result = u128::from_le_bytes(
-            self.source[self.pos..end_pos]
-                .try_into()
-                .map_err(|err| DeserializationError::UnknownError(format!("{err}")))?,
-        );
-
-        self.pos = end_pos;
-        Ok(result)
-    }
-
-    fn read_u8_vec(&mut self, len: usize) -> Result<Vec<u8>, DeserializationError> {
-        let end_pos = self.pos + len;
-        if end_pos > self.source.len() {
-            return Err(DeserializationError::UnexpectedEOF);
-        }
-        let result = self.source[self.pos..end_pos].to_vec();
-        self.pos = end_pos;
-        Ok(result)
-    }
-
-    fn read_u8_array<const N: usize>(&mut self) -> Result<[u8; N], DeserializationError> {
-        let end_pos = self.pos + N;
-        if end_pos > self.source.len() {
-            return Err(DeserializationError::UnexpectedEOF);
-        }
-        let result = self.source[self.pos..end_pos].try_into().map_err(|_| {
-            DeserializationError::UnknownError("failed to convert slide into an array".to_string())
-        })?;
-        self.pos = end_pos;
-        Ok(result)
-    }
-
-    fn has_more_bytes(&self) -> bool {
-        self.pos < self.source.len()
-    }
-}
-
-// BYTE WRITER
-// ================================================================================================
-
-/// Defines how primitive values are to be written into `Self`.
-pub trait ByteWriter: Sized {
-    // REQUIRED METHODS
-    // --------------------------------------------------------------------------------------------
-
-    /// Writes a single byte into `self`.
-    ///
-    /// # Panics
-    /// Panics if the byte could not be written into `self`.
-    fn write_u8(&mut self, value: u8);
-
-    /// Writes a sequence of bytes into `self`.
-    ///
-    /// # Panics
-    /// Panics if the sequence of bytes could not be written into `self`.
-    fn write_u8_slice(&mut self, values: &[u8]);
-
-    // PROVIDED METHODS
-    // --------------------------------------------------------------------------------------------
-
-    /// Writes a u16 value in little-endian byte order into `self`.
-    ///
-    /// # Panics
-    /// Panics if the value could not be written into `self`.
-    fn write_u16(&mut self, value: u16) {
-        self.write_u8_slice(&value.to_le_bytes());
-    }
-
-    /// Writes a u32 value in little-endian byte order into `self`.
-    ///
-    /// # Panics
-    /// Panics if the value could not be written into `self`.
-    fn write_u32(&mut self, value: u32) {
-        self.write_u8_slice(&value.to_le_bytes());
-    }
-
-    /// Writes a u64 value in little-endian byte order into `self`.
-    ///
-    /// # Panics
-    /// Panics if the value could not be written into `self`.
-    fn write_u64(&mut self, value: u64) {
-        self.write_u8_slice(&value.to_le_bytes());
-    }
-
-    /// Writes a serializable value into `self`.
-    ///
-    /// # Panics
-    /// Panics if the value could not be written into `self`.
-    fn write<S: Serializable>(&mut self, value: S) {
-        value.write_into(self)
-    }
-}
-
-impl ByteWriter for Vec<u8> {
-    fn write_u8(&mut self, value: u8) {
-        self.push(value);
-    }
-
-    fn write_u8_slice(&mut self, values: &[u8]) {
-        self.extend_from_slice(values);
-    }
-}
 
 // AS BYTES
 // ================================================================================================
